% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_bismark.R
\name{bismark_df_to_granges}
\alias{bismark_df_to_granges}
\title{Read Bismark output into a GRanges object}
\usage{
bismark_df_to_granges(bismark_df)
}
\description{
Read Bismark output into a GRanges object
}
