% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/granges_utils.R
\name{macs2_to_ranges_df}
\alias{macs2_to_ranges_df}
\title{Convert a MACS2 peaks file to a GRanges object}
\usage{
macs2_to_ranges_df(file_path, macs_file = "peaks_xls")
}
\arguments{
\item{file_path}{Path to the MACS2 peaks file.}

\item{macs_file}{The type of MACS2 file to read in.}
}
\description{
Convert a MACS2 peaks file to a GRanges object
}
