% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_bam.R
\name{merge_bam}
\alias{merge_bam}
\title{Merge BAM Files}
\usage{
merge_bam(bam_paths, output_path, env_name = "samtools-env")
}
\arguments{
\item{bam_paths}{Character Vector. Vector of paths to BAM files.}

\item{output_path}{Path to where to save files.}

\item{env_name}{The name of the Conda environment to use.}
}
\description{
Merge BAM Files
}
