% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_peaks_union.R
\name{merge_peaks_union}
\alias{merge_peaks_union}
\title{Union
Merge ChIP-Seq peaks by the union of the two peak sets.}
\usage{
merge_peaks_union(gr1, gr2)
}
\arguments{
\item{gr1}{The first GRanges object.}

\item{gr2}{The second GRanges object.}
}
\description{
Union
Merge ChIP-Seq peaks by the union of the two peak sets.
}
