% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_peaks.R
\name{plot_chip_peak_vs_cpg}
\alias{plot_chip_peak_vs_cpg}
\title{Plot Peaks
Plot ChIP peaks against CpG methylation frequency.}
\usage{
plot_chip_peak_vs_cpg(sum_df, experiment_name)
}
\arguments{
\item{sum_df}{The summary data frame.}

\item{experiment_name}{The name of the experiment. Used for the file name.}
}
\description{
Plot Peaks
Plot ChIP peaks against CpG methylation frequency.
}
