% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/granges_utils.R
\name{ranges_df_to_granges}
\alias{ranges_df_to_granges}
\title{Convert a data frame to a GRanges object}
\usage{
ranges_df_to_granges(df)
}
\arguments{
\item{df}{A data frame with columns chrom, chromStart, chromEnd, and other
columns to be kept in the GRanges object.}
}
\description{
Convert a data frame to a GRanges object
}
