% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_macs.R
\name{run_macs2}
\alias{run_macs2}
\title{Run MACS2 peak calling}
\usage{
run_macs2(
  sample_name,
  treatment_bam,
  control_bam,
  cutoff = 0.05,
  broad_cutoff = 0.1,
  broad = FALSE,
  env_name = "macs2-env"
)
}
\arguments{
\item{sample_name}{The name of the sample.}

\item{treatment_bam}{The path to the treatment BAM file.}

\item{control_bam}{The path to the control BAM file.}

\item{cutoff}{The cutoff to use for peak calling.}

\item{broad_cutoff}{The cutoff to use for broad peak calling.}

\item{broad}{Whether to call broad peaks.}

\item{env_name}{The name of the Conda environment to use.}
}
\description{
Run MACS2 peak calling
}
