% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_mapping.R
\name{run_mapping}
\alias{run_mapping}
\title{Run Mapping of FASTQ Files}
\usage{
run_mapping(
  sample_name,
  fastq_path,
  reference_fasta,
  env_name = "minimap2-env"
)
}
\arguments{
\item{sample_name}{The name of the sample.}

\item{fastq_path}{The path to the 1 or more FASTQ files. Those can be GZIP,
there is no need to uncompress them.}

\item{reference_fasta}{The path to the reference FASTA file,
the file can be GZIP compressed.}

\item{env_name}{The name of the Conda environment to use.}
}
\description{
Run mapping step for FASTQ files using \code{minimap2}.
Run minimap2 in the closest way possible to \code{bowtie2} implementation for
short reads.
Check \href{https://github.com/tyjo/coptr/issues/2}{this GitHub Issue} for
description of the comparison with \code{bowtie2}.
}
