% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_bam.R
\name{sort_bam}
\alias{sort_bam}
\title{Sort BAM Files}
\usage{
sort_bam(bam_path, env_name = "samtools-env")
}
\arguments{
\item{bam_path}{Character. Path to BAM file.}

\item{env_name}{The name of the Conda environment to use.}
}
\description{
Sort BAM Files
}
