% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum_df_to_gr.R
\name{sum_df_to_gr}
\alias{sum_df_to_gr}
\title{Convert a \code{sum_df} to a GRanges object}
\usage{
sum_df_to_gr(sum_df)
}
\arguments{
\item{sum_df}{A data frame with a region column.}
}
\description{
Convert a \code{sum_df} to a GRanges object.
Additional metadata is kept.
}
